C# libraries and example tested and in a shape to be released. However there is a bug in the aes220_ProgrammerGUI which still needs to be sorted.
Programmer GUI effectively broken. Configure and Program FPGA functions no longer working.
Non GUI programmer is fine so problem must be in GUI file.
Fixed a bug in the firmware hanging the micro-controller when erasing the FPGA. Was due to the RESET_F call in the firmware waiting for the startFpga() function to return TRUE which it doesn't if the Flash is blank...
Removed the conditional in the RESET_F call.