Modified the turn3v3On/Off functions so they write the status of the 3.3V rail to the EEPROM on the board.
The modification means that the board will now returns to the set 3.3V rail status on power up.
i.e. if the 3.3V rail is set to OFF then on of the first thing the micro-controller will do on power up will be to turn
it OFF. Note that this will prevent the FPGA from configuring if no 3.3V is provided externally.
However it allows for the board to be stacked up with only one board providing the 3.3V (or none if the 3.3V is provided
externally).

Sebastien Saury

29.08.2013